+++ TheTaleOfTheMysticalInspections/Player.h
@@ -7,28 +7,28 @@ using namespace std;
 namespace game {
 	class player : public entity {
 	private:
-		int strength;
-		int charisma;
-		int intelligence;
-		int constitution;
-		int wisdom;
-		int dexterity;
-		vector<int> allStats;
-		vector<int> setUpStats();
+		double strength;
+		double charisma;
+		double intelligence;
+		double constitution;
+		double wisdom;
+		double dexterity;
+		vector<double> allStats;
+		vector<double> setUpStats();
 	public:
-		int getStrength();
-		int getCharisma();
-		int getIntelligence();
-		int getConstitution();
-		int getWisdom();
-		int getDexterity();
-		vector<int> getAllStats();
-		void setStrength(int);
-		void setCharisma(int);
-		void setIntelligence(int);
-		void setConstitution(int);
-		void setWisdom(int);
-		void setDexterity(int);
-		void setAllStats(vector<int>);
+		double getStrength();
+		double getCharisma();
+		double getIntelligence();
+		double getConstitution();
+		double getWisdom();
+		double getDexterity();
+		vector<double> getAllStats();
+		void setStrength(double);
+		void setCharisma(double);
+		void setIntelligence(double);
+		void setConstitution(double);
+		void setWisdom(double);
+		void setDexterity(double);
+		void setAllStats(vector<double>);
 	};
 }
